Added km computation

+++ app/src/main/java/fr/legrand/runalytics/data/entity/RALocationDBEntity.kt
@@ -12,5 +12,5 @@ data class RALocationDBEntity(
     val altitude: Float = 0f,
     val fullAltitudeDiff: Float = 0f,
     val timestamp: Long = 0,
-    val currentKmTime: Int = 0
+    val currentKmTime: Long = 0
 )
